+++ packages/fluent-syntax-converters/.prettierrc
@@ -1,4 +1,4 @@
 {
-  "printWidth": 120,
+  "printWidth": 100,
   "singleQuote": true
 }
