+++ packages/fluent-syntax-converters/src/converters/fluent/converter.ts
@@ -1,9 +1,10 @@
-import { AdapterInterface } from './index';
+import { Converter } from '../../types/converter';
 import { Result } from '@inlang/common';
 import { parse, serialize, SingleResource } from '@inlang/fluent-syntax';
 
 /**
- * The fluent adapter exists for ease of use of the package.
+ * The fluent converter exists for ease of use of the package.
+ * In the sense that Fluent is a supported format.
  *
  * Under the hood, the adapter simply calls parse and serialize
  * as exposed by the fluent syntax package i.e. no format
@@ -15,9 +16,9 @@ import { parse, serialize, SingleResource } from '@inlang/fluent-syntax';
  * context of creating/editing translations. If the syntax is incorrect
  * the creator/editor of the translation should be aware of it.
  */
-export class FluentAdapter implements AdapterInterface {
-    parse(data: string): Result<SingleResource, Error> {
-        const parsedResource = parse(data, { withSpans: false });
+export class FluentConverter implements Converter {
+    parse(args: { data: string }): Result<SingleResource, Error> {
+        const parsedResource = parse(args.data, { withSpans: false });
         const junk = parsedResource.body.filter((entry) => entry.type === 'Junk');
         if (junk.length > 0) {
             return Result.err(Error("Couldn't parse the following entries:\n" + junk.map((junk) => junk.content)));
@@ -25,8 +26,8 @@ export class FluentAdapter implements AdapterInterface {
         return Result.ok(parsedResource);
     }
 
-    serialize(resource: SingleResource): Result<string, Error> {
-        const serialized = serialize(resource, { withJunk: false });
+    serialize(args: { resource: SingleResource }): Result<string, Error> {
+        const serialized = serialize(args.resource, { withJunk: false });
         return Result.ok(serialized);
     }
 }
