+++ packages/fluent-syntax-converters/src/converters/typesafe-i18n/converter.test.ts
@@ -1,30 +1,30 @@
 import { LanguageCode } from '@inlang/common';
-import { Typesafei18nAdapter } from '../src/typesafei18nAdapter';
+import { Typesafei18nConverter } from './converter';
 
-describe('Typesafei18nAdapter', () => {
-    const adapter = new Typesafei18nAdapter();
+describe('Typesafei18nConverter', () => {
+    const adapter = new Typesafei18nConverter();
     const serializationOptions = {
         isBaseLanguage: false,
         languageCode: 'de' as LanguageCode,
     };
     it('should parse ok', () => {
-        const result = adapter.parse(testFile);
+        const result = adapter.parse({ data: testFile });
         expect(result.isOk);
     });
 
     it('should serialize ok', () => {
-        const parse = adapter.parse(testFile);
+        const parse = adapter.parse({ data: testFile });
         if (parse.isErr) throw parse.error;
-        const serialize = adapter.serialize(parse.value, serializationOptions);
+        const serialize = adapter.serialize({ resource: parse.value }, serializationOptions);
         expect(serialize.isOk);
     });
 
     it('should have the same abstract syntax trees after multiple parsings and serializations', () => {
-        const parse1 = adapter.parse(testFile);
+        const parse1 = adapter.parse({ data: testFile });
         if (parse1.isErr) throw parse1.error;
-        const serialize1 = adapter.serialize(parse1.value, serializationOptions);
+        const serialize1 = adapter.serialize({ resource: parse1.value }, serializationOptions);
         if (serialize1.isErr) throw serialize1.error;
-        const parse2 = adapter.parse(serialize1.value);
+        const parse2 = adapter.parse({ data: serialize1.value });
         if (parse2.isErr) throw parse2.error;
         expect(parse1.value).toEqual(parse2.value);
     });
