+++ packages/fluent-syntax-converters/src/utils/parseResources.ts
@@ -1,6 +1,6 @@
 import { Result } from '@inlang/common';
 import { Resource, Resources } from '@inlang/fluent-syntax';
-import { AdapterInterface } from '../types/adapterInterface';
+import { Converter } from '../types/converter';
 import { SerializedResource } from '../types/serializedResource';
 
 /**
@@ -9,12 +9,12 @@ import { SerializedResource } from '../types/serializedResource';
  * The provided adapter determines from which file format.
  */
 export function parseResources(args: {
-    adapter: AdapterInterface;
+    converter: Converter;
     files: SerializedResource[];
 }): Result<Resources, Error> {
     const resources: Record<string, Resource | undefined> = {};
     for (const file of args.files) {
-        const parsed = args.adapter.parse(file.data);
+        const parsed = args.converter.parse({ data: file.data });
         if (parsed.isErr) {
             return Result.err(parsed.error);
         }
