+++ packages/fluent-syntax-converters/src/utils/serializeResources.test.ts
@@ -1,23 +1,29 @@
-import { parseResources } from '../../src/utils/parseResources';
-import { serializeResources } from '../../src/utils/serializeResources';
-import { adapters } from '../../src';
+import { parseResources } from './parseResources';
+import { serializeResources } from './serializeResources';
+import { converters } from '../converters/index';
 
 it('should serialize a file correctly', () => {
     const parsed = parseResources({
-        adapter: adapters.fluent,
+        converter: converters.fluent,
         files: [
             {
                 languageCode: 'en',
                 data: 'test = this is my test\nhello = hello there\ncomplex = Hello {$name}\nextra = a key without translations',
             },
-            { languageCode: 'da', data: 'test = dette er min test\nhello = hej med dig\ncomplex = Hej {$name}' },
-            { languageCode: 'de', data: 'test = dis ist ein test\nhello = hallo mit dich\ncomplex = Hallo {$name}' },
+            {
+                languageCode: 'da',
+                data: 'test = dette er min test\nhello = hej med dig\ncomplex = Hej {$name}',
+            },
+            {
+                languageCode: 'de',
+                data: 'test = dis ist ein test\nhello = hallo mit dich\ncomplex = Hallo {$name}',
+            },
         ],
     });
     if (parsed.isErr) {
         fail();
     }
-    const result = serializeResources({ adapter: adapters.fluent, resources: parsed.value });
+    const result = serializeResources({ converter: converters.fluent, resources: parsed.value });
     if (result.isErr) fail();
     expect(result.value).toEqual([
         {
