+++ packages/fluent-syntax-converters/src/utils/serializeResources.ts
@@ -1,6 +1,6 @@
 import { LanguageCode, Result } from '@inlang/common';
 import { Resources, SingleResource } from '@inlang/fluent-syntax';
-import { AdapterInterface } from '..';
+import { Converter } from '../types/converter';
 import { SerializedResource } from '../types/serializedResource';
 
 /**
@@ -9,12 +9,12 @@ import { SerializedResource } from '../types/serializedResource';
  * The provided adapter determines to which file format.
  */
 export function serializeResources(args: {
-    adapter: AdapterInterface;
+    converter: Converter;
     resources: Resources;
 }): Result<SerializedResource[], Error> {
     const files: SerializedResource[] = [];
     for (const [languageCode, resource] of Object.entries(args.resources.resources)) {
-        const serialized = args.adapter.serialize(resource as SingleResource);
+        const serialized = args.converter.serialize({ resource: resource as SingleResource });
         if (serialized.isErr) {
             return Result.err(serialized.error);
         }
